refactor: move package to root and rename to book_to_skill
