+++ book_to_skill/__main__.py
@@ -1,0 +1,4 @@
+from book_to_skill.cli import main
+
+if __name__ == "__main__":
+    main()
