+++ book_to_skill/cli.py
@@ -1,0 +1,15 @@
+import sys
+from book_to_skill.utils import main as utils_main
+
+def main():
+    # Force UTF-8 stdout/stderr to avoid UnicodeEncodeError on Windows console
+    for _stream in (sys.stdout, sys.stderr):
+        try:
+            _stream.reconfigure(encoding="utf-8")
+        except (AttributeError, ValueError):
+            pass
+    utils_main()
+
+# Expose main for packaging console scripts entry points
+if __name__ == "__main__":
+    main()
