+++ book_to_skill/dependencies.py
@@ -5,7 +5,7 @@ import os
 import shutil
 import subprocess
 import sys
-from extractor.config import PYTHON_DEPENDENCIES, HTML_EXTENSIONS
+from book_to_skill.config import PYTHON_DEPENDENCIES, HTML_EXTENSIONS
 
 
 # Ordered groups for the --check preflight report. Each entry describes one
