+++ book_to_skill/exceptions.py
