--- scripts/extractor/__init__.py
@@ -1,4 +1,0 @@
-from extractor.utils import resolve_input_files, extract_single_file, main
-from extractor.exceptions import ExtractionError
-
-__all__ = ["resolve_input_files", "extract_single_file", "main", "ExtractionError"]
